+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.2.3
 - Fixed fluid cache updating wrongly (raoulvdberge)
+- Fixed Exporter scheduling too many tasks (raoulvdberge)
 
 ### 1.2.2
 - Various autocrafting fixes (way2muchnoise)